Split walk code
